+++ Assets/Script/GameScene/Player/PlayerControl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb68a0cc07e8bcd40a21db1cf0b4820b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
